splitting networks into backbones and heads, initial commit
